% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_ggsurvfit.R
\name{kmfun_ggsurvfit}
\alias{kmfun_ggsurvfit}
\title{km_ggsurvfit}
\usage{
kmfun_ggsurvfit(
  mytime,
  entry = "traditional",
  mytime_t0_to_entry = NULL,
  mydelta,
  group = NULL,
  mydata,
  myxlab = "Specify x axis label",
  myylab = "Survival probability",
  set_linetype_aes = FALSE,
  legendtitle = NULL,
  mytitle = NULL,
  mysubtitle = NULL,
  maxtimeset = NULL,
  xaxis_max_adjust = 2,
  coord_cartesian_xmax = NULL,
  xbreaksby = 12,
  print_plot = TRUE,
  test = 0,
  risk.table.set = TRUE,
  conf.int.set = TRUE
)
}
\arguments{
\item{mytime}{(quoted) Survival time variable}

\item{entry}{(quoted) What type of risk set entry are we using? "traditional" (default) or "delayed" (implements entry risk set adjustment approach)}

\item{mytime_t0_to_entry}{(quoted) Variable for time from baseline to risk set entry (only use when implementing entry risk set adjustment)}

\item{mydelta}{(quoted) Survival status variable. Must be numeric with: \code{0} = censor, \code{1} = event of interest}

\item{group}{(quoted) Group variable for plotting groups}

\item{mydata}{(unquoted) A data frame}

\item{myxlab}{(quoted) x-axis label text. If \code{NULL}, then this will default to \code{labelled::var_label(mydata[[mytime]])} if \code{!is.null(labelled::var_label(mydata[[mytime]]))} = \code{TRUE}}

\item{myylab}{(quoted) y-axis label text}

\item{set_linetype_aes}{Logical for whether to use \code{ggplot2::scale_linetype_manual()}}

\item{legendtitle}{(quoted) Legend title text}

\item{mytitle}{(quoted) Title text}

\item{mysubtitle}{(quoted) Subtitle text}

\item{maxtimeset}{Numeric value for max value on x-axis. If \code{NULL}, will default to the maximum of the \code{mytime} arg}

\item{xaxis_max_adjust}{Numeric value for adding a buffer to the max value on x-axis}

\item{coord_cartesian_xmax}{Numeric value for clipping the x-axis at a max value}

\item{xbreaksby}{Numeric value for controlling x-axis breaks}

\item{print_plot}{Logical for whether to print the plot that gets returned. Accepts \code{TRUE} or \code{FALSE}}

\item{test}{Logical for whether to include a Log-Rank p-value on the plot. Accepts \code{0} or \code{1}}

\item{risk.table.set}{Logical for whether to add a risk table. Accepts \code{TRUE} or \code{FALSE}}

\item{conf.int.set}{Logical for whether to add confidence bands. Accepts \code{TRUE} or \code{FALSE}}
}
\value{
survplot
}
\description{
Returns a Kaplan-Meier curve created using ggsurvfit
}
\seealso{
\code{\link[ggsurvfit]{ggsurvfit}}
}
