% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_readwrite.R
\name{mod_readwrite_ui}
\alias{mod_readwrite_ui}
\alias{mod_readwrite_server}
\title{mod_readwrite_ui and mod_readwrite_server}
\usage{
mod_readwrite_ui(id)

mod_readwrite_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
