% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_slobr.R
\name{mod_slobr_ui}
\alias{mod_slobr_ui}
\alias{mod_slobr_server}
\title{mod_slobr_ui and mod_slobr_server}
\usage{
mod_slobr_ui(id)

mod_slobr_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
