% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_slobr}
\alias{run_slobr}
\title{Run the Shiny Application}
\usage{
run_slobr(conn = "demo")
}
\arguments{
\item{conn}{A \linkS4class{SQLiteConnection} to a database or 'demo' to open with slobr demo database.}
}
\description{
Run the Shiny Application
}
