% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slobR-package.R
\docType{package}
\name{slobR-package}
\alias{slobR}
\alias{slobR-package}
\title{slobR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
