% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_x_by_y.R
\name{viz_x_by_y}
\alias{viz_x_by_y}
\title{viz_x_by_y}
\usage{
viz_x_by_y(mydata, x, facet_grid_col_var)
}
\arguments{
\item{mydata}{(unquoted) A data frame}

\item{x}{(quoted) A variable name for the x-axis}

\item{facet_grid_col_var}{(quoted) A categorical variable for the facet_grid column variable}
}
\value{
plot
}
\description{
Visualize one var (continuous or categorical) by one facet grid column variable (categorical).
}
