% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(path = NULL)
}
\arguments{
\item{path}{A path to SQLite database, or "demo" to load demo database, or NULL
to load the app without a database.}
}
\description{
Run the Shiny Application
}
